Added photos
I added in the the photos for the website and made a few style adjustments to the links. I also need access to the google docs to add to the website. Alok, the spacing of the elements is a bit off, could you take a look at that with me? I also need help centering the art titles vertically. Thanks!